% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aln_rates.R
\name{aln_metrics<-}
\alias{aln_metrics<-}
\title{Setter for aln_metrics slot}
\usage{
aln_metrics(x) <- value
}
\arguments{
\item{x}{A BbcSE object.}

\item{value}{A matrix of mapping metrics. Rownames must correspond to samples
in counts matrix and column names indicating metric type must be set.}
}
\description{
'map_rate' and 'uniq_map_rate' will be calculated if not already present.
}
