% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aln_rates.R
\name{aln_metrics}
\alias{aln_metrics}
\title{Getter for aln_metrics slot}
\usage{
aln_metrics(x, withDimnames = TRUE)
}
\arguments{
\item{x}{BbcSE object}

\item{withDimnames}{logical indicating whether output should have dim names}
}
\description{
Getter for aln_metrics slot
}
