% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aln_rates.R
\docType{methods}
\name{aln_rates,BbcSE-method}
\alias{aln_rates,BbcSE-method}
\title{Getter for aln_rates slot}
\usage{
\S4method{aln_rates}{BbcSE}(x, withDimnames = TRUE)
}
\description{
Getter for aln_rates slot
}
