% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aln_rates.R
\docType{methods}
\name{aln_rates<-,BbcSE-method}
\alias{aln_rates<-,BbcSE-method}
\title{Setter for aln_rates slot}
\usage{
\S4method{aln_rates}{BbcSE}(x) <- value
}
\arguments{
\item{x}{A BbcSE object.}

\item{value}{A matrix of mapping metrics. Rownames must correspond to samples
in counts matrix and column names indicating metric type must be set.}
}
\description{
Setter for aln_rates slot
}
