% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aln_rates.R
\name{aln_rates}
\alias{aln_rates}
\title{Getter for aln_rates slot}
\usage{
aln_rates(x, withDimnames = TRUE)
}
\arguments{
\item{x}{BbcSE object}

\item{withDimnames}{logical indicating whether output should have dim names}
}
\description{
Getter for aln_rates slot
}
