% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{batch_correct_norm_cts}
\alias{batch_correct_norm_cts}
\title{Batch correct normalized counts}
\usage{
batch_correct_norm_cts(
  x,
  de_method = "edger",
  correction_method = "removeBatchEffect",
  new_assay_name = "batch_corr",
  ...
)
}
\arguments{
\item{x}{A BbcSE object}

\item{de_method}{"edger" or "deseq2"}

\item{correction_method}{"removeBatchEffect" or "combat"}

\item{new_assay_name}{Name of the batch-corrected counts stored in new assay
in 'norm_cts' slot.}

\item{...}{passed to batch correction function.}
}
\value{
A BbcSE object
}
\description{
Batch correct normalized counts for visualizations or other downstream non-DE
analysis applications.
}
\examples{
\dontrun{
# Default is limma::removeBatchEffect
bbc_obj_batch <- batch_correct_norm_cts(bbc_obj, batch=colData(bbc_obj)$Rep,
design=model.matrix(~Condition, data=colData(bbc_obj)))

plot_PCA(bbc_obj_batch, assay_name = "batch_corr", adonis=FALSE,
color_by="Time", shape_by="Rep")

# Combat is also supported
bbc_obj_batch <- batch_correct_norm_cts(bbc_obj, new_assay_name="combat",
batch=colData(bbc_obj)$Rep, correction_method = "combat", mod =
model.matrix(~Condition, data=colData(bbc_obj)))

plot_PCA(bbc_obj_batch, assay_name = "combat", adonis=FALSE, color_by="Time",
shape_by="Rep")
}
}
\seealso{
\code{\link[limma]{removeBatchEffect} \link[sva]{ComBat}}
}
