% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BbcEdgeR_manip_methods.R
\name{BbcEdgeR}
\alias{BbcEdgeR}
\title{Constructor for BbcEdgeR}
\usage{
BbcEdgeR(dgelist = edgeR::DGEList(), de_results = list(),
  norm_cts = SummarizedExperiment())
}
\arguments{
\item{dgelist}{A DGEList}

\item{de_results}{A list containing DGEGLM as first element and then edgeR
result objects}

\item{norm_cts}{A SummarizedExperiment}
}
\value{
A BbcEdgeR object.
}
\description{
\code{BbcEdgeR} is a constructor for BbcEdgeR
}
