% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbcRNA.R
\docType{package}
\name{bbcRNA}
\alias{bbcRNA}
\alias{bbcRNA-package}
\title{bbcRNA: Bulk RNA-seq workflow for VAI BBC..}
\description{
This package provides functions and a class (bbcRNA) to perform
standard bulk RNA-seq analyses.
}
\section{Functions}{

star_to_mat
}

