% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-classes.R
\docType{class}
\name{BbcRNAData-class}
\alias{BbcRNAData-class}
\alias{.BbcRNAData}
\title{BbcRNAData, an S4 class to represent RNA-seq data in various containers}
\description{
BbcRNAData, an S4 class to represent RNA-seq data in various containers
}
\section{Slots}{

\describe{
\item{\code{summexp}}{A RangedSummarizedExperiment.}

\item{\code{edger}}{A list containing DGEList (must be first element) and edgeR
results objects.}

\item{\code{deseq2}}{A DESeqDataSet object.}
}}

