% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-classes.R
\name{BbcRNAData}
\alias{BbcRNAData}
\title{Constructor for BbcRNAData}
\usage{
BbcRNAData(counts, ...)
}
\arguments{
\item{counts}{A count matrix with sample names as column names and gene names
as row names.}

\item{...}{Arguments for SummarizedExperiment.}
}
\description{
\code{BbcRNAData} is a constructor for BbcRNAData.
}
