% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-classes.R
\name{BbcSE}
\alias{BbcSE}
\title{Constructor for BbcSE}
\usage{
BbcSE(counts, ...)
}
\arguments{
\item{counts}{A count matrix with sample names as column names and gene names
as row names.}

\item{...}{Arguments for SummarizedExperiment.}
}
\description{
\code{BbcSE} is a constructor for BbcSE.
}
