% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BbcSE_manip_methods.R
\name{cbind,BbcSE-method}
\alias{cbind,BbcSE-method}
\title{Column combining method for BbcSE}
\usage{
\S4method{cbind}{BbcSE}(..., deparse.level = 1)
}
\arguments{
\item{...}{BbcSE objects}

\item{deparse.level}{See ?base::cbind for a description of this argument.}
}
\description{
Column combining method for BbcSE
}
