% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BbcEdgeR_manip_methods.R
\name{de_results<-}
\alias{de_results<-}
\title{Setter for de_results slot in a BbcEdgeR object}
\usage{
de_results(x) <- value
}
\arguments{
\item{x}{A BbcEdgeR object.}

\item{value}{A list containing DGEGLM in first element and then edgeR result
objects}
}
\description{
Setter for de_results slot in a BbcEdgeR object
}
