% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BbcEdgeR_manip_methods.R
\name{de_results}
\alias{de_results}
\title{Getter for de_results slot from BbcEdgeR object}
\usage{
de_results(x)
}
\arguments{
\item{x}{BbcEdgeR object}
}
\description{
Getter for de_results slot from BbcEdgeR object
}
