% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BbcEdgeR_manip_methods.R
\name{dgelist<-}
\alias{dgelist<-}
\title{Setter for dgelist slot in a BbcEdgeR object}
\usage{
dgelist(x) <- value
}
\arguments{
\item{x}{A BbcEdgeR object.}

\item{value}{A DGEList object}
}
\description{
Setter for dgelist slot in a BbcEdgeR object
}
