% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BbcEdgeR_manip_methods.R
\name{dgelist}
\alias{dgelist}
\title{Getter for dgelist slot from BbcEdgeR object}
\usage{
dgelist(x)
}
\arguments{
\item{x}{BbcEdgeR object}
}
\description{
Getter for dgelist slot from BbcEdgeR object
}
