% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_data_methods.R
\name{ens2entrez}
\alias{ens2entrez}
\title{Convert Ensembl IDs to Entrez IDs for gene enrichment analyses.}
\usage{
ens2entrez(x, orgdb, BMdataset = "")
}
\arguments{
\item{x}{A BbcSE object.}

\item{orgdb}{A OrgDb object. Download the annotations for your species from
'http://bioconductor.org/packages/release/BiocViews.html#___OrgDb'. Only
used if BMdataset="".}

\item{BMdataset}{character. The name of the BM dataset you want to use. Try
to use 'bbcRNA::searchBM("species_name")' to find the appropriate dataset
if needed.}
}
\value{
A BbcSE object.
}
\description{
Convert Ensembl IDs to Entrez IDs for gene enrichment analyses.
}
\details{
Entrez IDs with more than one Ensembl ID match are determined only from genes
present in the BbcSE object, so be sure it is unfiltered.

\itemize{
\item Entrez IDs that match multiple Ensembl IDs are NA
\item The first Entrez ID will be returned for genes with multiple possible
Entrez IDs.
\item Genes with no Entrez ID or absent from the database will be
NA. }
}
