% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_data_methods.R
\docType{methods}
\name{ens2sym}
\alias{ens2sym}
\title{Convert Ensembl IDs to gene symbols.}
\usage{
\S4method{ens2sym}{BbcSE}(x, orgdb)
}
\arguments{
\item{x}{A BbcSE object.}

\item{orgdb}{A OrgDb object. Download the annotations for your species from
'http://bioconductor.org/packages/release/BiocViews.html#___OrgDb'}
}
\value{
A BbcSE object.
}
\description{
Convert Ensembl IDs to gene symbols.
}
\details{
Non 1:1 matches are determined from all the Ensembl IDs in the OrgDb, not
just those in the BbcSE object.
\itemize{
  \item Symbols that match multiple Ensembl IDs will be resolved by concatenating
the Ensembl ID and the gene symbol. Uses scater::uniquifyFeatureNames.
  \item Genes with multiple possible symbols will be labelled as NA.
  \item Genes absent from OrgDb will be labelled as Ensembl ID.
}
}
