% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea.R
\name{find_missing_in_gseaResult}
\alias{find_missing_in_gseaResult}
\title{Compare genes in clusterProfiler gseaResult to genes in a bbcRNA object}
\usage{
find_missing_in_gseaResult(gseaResults, bbcRNA, de_pkg = "edger")
}
\arguments{
\item{gseaResults}{A list of ClusterProfiler::gseaResult objects outputted by 'run_gsea'.}

\item{bbcRNA}{bbcRNA object}

\item{de_pkg}{"edger" or "deseq2"}
}
\value{
A dataframe of missing genes.
}
\description{
Output dataframe for genes that did not have an Entrez match (based on rules
in ens2entrez) and genes that are present in the bbcRNA object but not in the
bbcRNA object. Extracts LFC and PValue from DE results. If present in the
bbcRNA object but not in the DE analysis, then it is inferred to be a lowly
expressed gene. Includes all rowData.
}
