% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_output.R
\name{get_de_table}
\alias{get_de_table}
\title{Get a table of DE results}
\usage{
get_de_table(x, de_pkg = "edger")
}
\arguments{
\item{x}{A BbcSE object}

\item{de_pkg}{"edger" or "deseq2"}
}
\value{
A list of datafrane for each contrast.
}
\description{
Get a table of DE results
}
