% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edger_methods.R
\name{lists_to_edger_contrasts}
\alias{lists_to_edger_contrasts}
\title{Function to convert to edgeR-style contrasts}
\usage{
lists_to_edger_contrasts(contrasts_list, design)
}
\arguments{
\item{contrasts_list}{list of chr vectors containing variable name, numerator
level, denominator level}

\item{design}{output of model.matrix}
}
\description{
Function to convert to edgeR-style contrasts
}
