% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BbcEdgeR_manip_methods.R
\name{norm_cts<-}
\alias{norm_cts<-}
\title{Setter for norm_cts slot in a BbcEdgeR object}
\usage{
norm_cts(x) <- value
}
\arguments{
\item{x}{A BbcEdgeR object.}

\item{value}{A list containing DGEGLM in first element and then edgeR result
objects}
}
\description{
Setter for norm_cts slot in a BbcEdgeR object
}
