% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BbcEdgeR_manip_methods.R
\name{norm_cts}
\alias{norm_cts}
\title{Getter for norm_cts slot from BbcEdgeR object}
\usage{
norm_cts(x)
}
\arguments{
\item{x}{BbcEdgeR object}
}
\description{
Getter for norm_cts slot from BbcEdgeR object
}
