% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/edger_methods.R
\docType{methods}
\name{normalize_counts}
\alias{normalize_counts}
\alias{normalize_counts,BbcSE-method}
\alias{normalize_counts,DGEList-method}
\title{Calculate normalized counts.}
\usage{
normalize_counts(x, ...)

\S4method{normalize_counts}{BbcSE}(x, de_pkg = "edger")

\S4method{normalize_counts}{DGEList}(x)
}
\arguments{
\item{x}{A BbcSE object or a DGEList or a DESeqDataSet.}

\item{...}{Not used currently}

\item{de_pkg}{"edger" or "deseq2". Only used if x is a BbcSE}
}
\value{
A BbcSE object or a matrix or....
}
\description{
Calculate normalized counts.
}
\details{
Normalization uses methods from the package corresponding to the de_pkg
paramter:
\describe{ \item{edger}{Uses edgeR::cpm with normalized.lib.sizes =
TRUE and log = TRUE}
\item{deseq2}{Not implemented yet.} }
}
\section{Methods (by class)}{
\itemize{
\item \code{BbcSE}: For de_pkg="edger", a SummarizedExperiment
object is created and stored in the BbcEdgeR object in the edger slot.
Group average log normalized counts are also calculated and stored as
rowData.

\item \code{DGEList}: log normalized counts returned as a matrix.
}}

\seealso{
\code{\link[edgeR]{cpm}}
}
