% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aln_rates.R
\name{plot_aln_metrics}
\alias{plot_aln_metrics}
\title{Plot alignment metrics}
\usage{
plot_aln_metrics(x, type = "uniq_map_rate", fill_by = "sample",
  facet_by = "")
}
\arguments{
\item{x}{A BbcSE object.}

\item{type}{aln_metrics column to plot.}

\item{fill_by}{Variable to fill by. Can be column from colData() or
aln_metrics().}

\item{facet_by}{Variable to facet by. Can be column from colData() or
aln_metrics().}
}
\value{
A ggplot object
}
\description{
Plot alignment metrics as a bar chart.
}
