% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aln_rates.R
\docType{methods}
\name{plot_aln_rates}
\alias{plot_aln_rates}
\title{Plot alignment metrics}
\usage{
\S4method{plot_aln_rates}{BbcSE}(x, type, fill_by = "sample",
  facet_by = "")
}
\arguments{
\item{x}{A BbcSE object.}

\item{type}{aln_rates column to plot.}

\item{fill_by}{Variable to fill by. Can be column from colData() or
aln_data().}

\item{facet_by}{Variable to facet by. Can be column from colData() or
aln_data().}
}
\value{
A ggplot object
}
\description{
Plot alignment metrics as a bar chart.
}
