% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Plot heatmap}
\usage{
plot_heatmap(x, genes = rownames(x), de_method = "edger",
  assay_name = "norm_log_cpm", gene_labels = NULL,
  coldata_annot = NULL, coldata_split = NULL, rowdata_annot = NULL,
  rowdata_split = NULL, clust_rows = TRUE, clust_cols = TRUE,
  zscores = FALSE, grouped = FALSE)
}
\arguments{
\item{x}{A BbcSE object}

\item{genes}{gene IDs (rownames)}

\item{de_method}{"edger" or "deseq2"}

\item{assay_name}{Name of assay from the 'norm_cts' slot. Not compatible with
grouped=TRUE. Option was implemented with the idea of being able to plot
batch-corrected data.}

\item{gene_labels}{a column name from rowData. Must contain unique values}

\item{coldata_annot}{character vector of colData colnames for annotating}

\item{coldata_split}{character value of colData colname for splitting}

\item{rowdata_annot}{character vector of rowData colnames for annotating}

\item{rowdata_split}{character value of rowData colname for splitting}

\item{clust_rows}{logical for whether rows should be clustered}

\item{clust_cols}{logical for whether columns should be clustered}

\item{zscores}{logical indicating whether the expression matrix should be
converted to Z-scores}

\item{grouped}{logical indicating whether the mean expression per group
should be shown.}
}
\value{
A Heatmap-class object.
}
\description{
Plot heatmap with option for splitting rows and columns and adding
annotation.
}
\seealso{
\code{\link[ComplexHeatmap]{Heatmap}}
  \href{https://jokergoo.github.io/ComplexHeatmap-reference/book/}{ComplexHeatmap
   book}
}
