% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{plot_PCA}
\alias{plot_PCA}
\title{PCA, screeplot and PERMANOVA test}
\usage{
plot_PCA(x, norm_cts_type = "edger",
  color_by = colnames(colData(x))[[1]],
  shape_by = colnames(colData(x))[[1]], adonis = TRUE,
  adonis_by = colnames(colData(x))[[1]])
}
\arguments{
\item{x}{A BbcSE object.}

\item{norm_cts_type}{"edger" or "deseq2"}

\item{color_by}{colData column to color by for PCA plot}

\item{shape_by}{colData column to shape by for PCA plot}

\item{adonis}{logical for whether vegan::adonis should be run. If TRUE,
adonis() will be run with Euclidean distance calculated from the same
normalized counts as used for PCA.}

\item{adonis_by}{colData column to test using PERMANOVA. May be a vector of
values; in this case, each variable is tested sequentially using adonis().}
}
\value{
A list containing ggplot objects for 1. PCA 2. scree plot 3. a list
  of output from vegan::adonis()
}
\description{
Runs PCA on normalized counts. Makes PCA and scree plots. Tests for
significant clustering based on specific variables using PERMANOVA and
Euclidean distances.
}
\seealso{
\code{\link[vegan]{adonis}}
}
