% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{plot_pval_distrib}
\alias{plot_pval_distrib}
\title{Plot P-values}
\usage{
plot_pval_distrib(x, de_method = "edger", contrast_names)
}
\arguments{
\item{x}{A BbcSE object}

\item{de_method}{"edger" or "deseq2"}

\item{contrast_names}{character value or vector for the contrast(s) of
interest. See name(de_results(edger(x))). If not specified, then all
contrasts will be processed.}
}
\value{
A ggplot object
}
\description{
Plot P-value distribution for contrasts in edger or deseq2 slot of BbcSE
object
}
