% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{plot_volcano}
\alias{plot_volcano}
\title{Volcano plot}
\usage{
plot_volcano(
  x,
  de_method = "edger",
  gene_name = "gene_symbol",
  pval_name = "FDR",
  pvalCutoff = 0.05,
  y_title = paste0("-Log10 (", pval_name, ")"),
  contrast_names,
  ...
)
}
\arguments{
\item{x}{A BbcSE object}

\item{de_method}{"edger" or "deseq2"}

\item{gene_name}{"gene_symbol" or "ensembl_id". Corresponds to columns in
'bbcRNA::get_de_table()' output}

\item{pval_name}{"PValue" or "FDR". Corresponds to columns in
'bbcRNA::get_de_table()' output}

\item{pvalCutoff}{"Cut-off for statistical significance. A horizontal line
will be drawn at -log10(pCutoff)." See EnhancedVolcano documentation.}

\item{y_title}{Y axis label}

\item{contrast_names}{character value or vector for the contrast(s) of
interest. See name(de_results(edger(x))). If not specified, then all
contrasts will be processed.}

\item{...}{Passed to EnhancedVolcano::EnhancedVolcano}
}
\value{
A list of ggplot objects
}
\description{
Plot volcano plots for contrasts in edger or deseq2 slot of BbcSE object
}
\seealso{
\code{\link[EnhancedVolcano]{EnhancedVolcano}}
}
