% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea.R
\name{prep_clusterprofiler_genelist}
\alias{prep_clusterprofiler_genelist}
\title{Prep clusterProfiler geneList}
\usage{
prep_clusterprofiler_genelist(gene_and_metric_df)
}
\arguments{
\item{gene_and_metric_df}{two column df with Entrez ID as col1 and
metric(log2FC or F etc) as col2}
}
\description{
Prep clusterProfiler geneList
Adapts code from
https://github.com/YuLab-SMU/DOSE/wiki/how-to-prepare-your-own-geneList.
}
