% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{methods}
\name{rbind,BbcSE-method}
\alias{rbind,BbcSE-method}
\title{Row combining method for BbcSE}
\usage{
\S4method{rbind}{BbcSE}(..., deparse.level = 1)
}
\description{
Row combining method for BbcSE
}
