% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_input.R
\name{read_col_meta}
\alias{read_col_meta}
\title{Read column meta data}
\usage{
read_col_meta(file)
}
\arguments{
\item{file}{A tab-delimited column meta data file. Must have a column named
"sample".}
}
\value{
A DataFrame with 'sample' column set as row names.
}
\description{
\code{read_col_meta} returns a dataframe for adding to a BbcSE object.
}
\details{
This function will read in column meta data.
}
