% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_input.R
\name{read_star_aln_metrics}
\alias{read_star_aln_metrics}
\title{Import STAR Log.final.out files for analyzing mapping rates.}
\usage{
read_star_aln_metrics(dir, rgx)
}
\arguments{
\item{dir}{A character scalar indicating the directory containing all the
STAR Log.final.out files.}

\item{rgx}{A character scalar representing a regex used to parse out the
sample name from the name of the Log.final.out files.}
}
\value{
A dataframe.
}
\description{
\code{read_star_aln_metrics} returns a dataframe from STAR Log.final.out files.
}
\details{
This function will read in all the '*Log.final.out' files in the specified
directory and returns a dataframe with alignment metrics.
}
