% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/gsea.R
\docType{methods}
\name{run_gsea}
\alias{run_gsea}
\alias{run_gsea,data.frame-method}
\alias{run_gsea,BbcSE-method}
\title{Run GSEA}
\usage{
run_gsea(x, ...)

\S4method{run_gsea}{data.frame}(x, gene_set = "reactome", organism,
  orgDb, ...)

\S4method{run_gsea}{BbcSE}(x, de_pkg = "edger",
  rank_by = "signed-log10pval", contrast_names = "", ...)
}
\arguments{
\item{x}{A BbcSE object or a dataframe.}

\item{...}{Passed to gseKEGG, GSEA or gsePathway. See each function for
possible arguments.}

\item{gene_set}{one of "kegg", "reactome", "H", "C1", "C2", "C3", "C4", "C5",
"C6", "C7"}

\item{organism}{For compatibility with dowstream tools, the organism is
specified differently depending on the desired gene set. For KEGG, use the
three-letter code according to
http://www.genome.jp/kegg/catalog/org_list.html. For Reactome, possible
values are ‘celegans’, ‘fly’, ‘human’, ‘mouse’, ‘rat’, ‘yeast’ and
‘zebrafish’. For msigdb, run 'msigdbr::msigdbr_show_species()'.}

\item{orgDb}{OrgDB object for your organism}

\item{de_pkg}{"edger" or "deseq2". Only used if x is a BbcSE}

\item{rank_by}{\describe{
 \item{"signed-log10pval"}{-log10 of the PValue from the DE analysis, signed
 by the LFC.}
 \item{"-log10pval"}{-log10 of the PValue from the DE analysis. May be ideal
 for gene sets that contain both up-regulated and down-regulated genes for a
 particular process or pathway etc.}
}}

\item{contrast_names}{character value or vector for the contrast(s) of
interest. See name(de_results(edger(x))). If left to default "" value then
all contrasts will be processed.}
}
\value{
A list of gseaResult objects or one gseaResult object
}
\description{
Run GSEA
}
\section{Methods (by class)}{
\itemize{
\item \code{data.frame}: x is a dataframe with gene name as col1 and
metric(log2FC or F etc) as col2. Function will sort by col2 in decreasing
order. See '?prep_clusterprofiler_genelist'.

\item \code{BbcSE}: Extract out the gene and signed -log10 PValue. Run
run_gsea for each result object (contrast).
}}

\seealso{
\code{\link[clusterProfiler]{gseKEGG} \link[clusterProfiler]{GSEA}
\link[ReactomePA]{gsePathway}}
}
