% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/gsea.R
\name{run_hypergeometric}
\alias{run_hypergeometric}
\alias{run_hypergeometric,data.frame-method}
\alias{run_hypergeometric,BbcSE-method}
\title{Run hypergeometric test}
\usage{
run_hypergeometric(x, ...)

\S4method{run_hypergeometric}{data.frame}(
  x,
  gene_set = "reactome",
  organism,
  orgDb,
  split_by_lfc_dir = TRUE,
  padj_cutoff = 0.05,
  lfc_cutoff = 0,
  ...
)

\S4method{run_hypergeometric}{BbcSE}(
  x,
  de_pkg = "edger",
  contrast_names = "",
  split_by_lfc_dir = TRUE,
  ...
)
}
\arguments{
\item{x}{A BbcSE object.}

\item{...}{Passed to enrichKEGG, enrichGO, enrichPathway or enricher. See
each function for possible arguments.}

\item{gene_set}{one of "kegg", "reactome", "H", "C1", "C2", "C3", "C4", "C5",
"C6", "C7"}

\item{organism}{For compatibility with dowstream tools, the organism is
specified differently depending on the desired gene set. For KEGG, use the
three-letter code according to
http://www.genome.jp/kegg/catalog/org_list.html. For Reactome, possible
values are ‘celegans’, ‘fly’, ‘human’, ‘mouse’, ‘rat’, ‘yeast’ and
‘zebrafish’. For msigdb, run 'msigdbr::msigdbr_show_species()'.}

\item{orgDb}{OrgDB object for your organism}

\item{split_by_lfc_dir}{logical. Whether DE genes should be split into up and
down-regulated}

\item{padj_cutoff}{numeric. Cutoff for adjusted PValue for DE.}

\item{lfc_cutoff}{numeric. Cutoff for LFC (absolute value) for DE.}

\item{de_pkg}{"edger" or "deseq2". Only used if x is a BbcSE}

\item{contrast_names}{character value or vector for the contrast(s) of
interest. See name(de_results(edger(x))). If left to default "" value then
all contrasts will be processed.}
}
\value{
A list of gseaResult objects or one gseaResult object
}
\description{
Run hypergeometric test
}
\section{Methods (by class)}{
\itemize{
\item \code{data.frame}: x is a dataframe with gene name (entrez) as
col1, log2FC as col2 and adjusted PValue as col3. Function will filter by
LFC and PValue for DE genes, and optionally will split the DE genes into up
and down-regulated based on LFC. The gene universe are all the genes in the
dataframe. Returns a list of enrichResults for up and down-regulaed genes
or just all the DE genes without regard for direction.

\item \code{BbcSE}: Extract out the gene, LFC and adjusted PValue.
Run run_hypergeometric for each result object (contrast).
}}

\seealso{
\code{\link[clusterProfiler]{enrichKEGG}
  \link[clusterProfiler]{enrichGO} \link[ReactomePA]{enrichPathway}
  \link[clusterProfiler]{enricher}}
}
