% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_data_methods.R
\name{searchBM}
\alias{searchBM}
\title{Search biomart datasets with keyword}
\usage{
searchBM(keyword)
}
\arguments{
\item{keyword}{regex to search for. Perl style.}
}
\description{
Search biomart datasets with keyword
}
