% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BbcSE_manip_methods.R
\docType{methods}
\name{show,BbcSE-method}
\alias{show,BbcSE-method}
\title{show method for BbcSE}
\usage{
\S4method{show}{BbcSE}(object)
}
\arguments{
\item{object}{BbcSE object}
}
\description{
show method for BbcSE
}
