% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_input.R
\name{star_to_mat}
\alias{star_to_mat}
\title{Convert STAR read counts to count matrix.}
\usage{
star_to_mat(dir, rgx, column)
}
\arguments{
\item{dir}{A character scalar indicating the directory containing all the
STAR ReadsPerGene.out.tab files.}

\item{rgx}{A character scalar representing a regex used to parse out the
sample name from the name of the ReadsPerGene.out.tab file.}

\item{column}{An integer indicating the column to extract counts from.
1 = unstranded, 2 = 1st read strand aligned with RNA, 3 = 2nd read
strand aligned with RNA  with RNA}
}
\value{
A matrix where the column names are the sample names and the row
    names are the gene names
}
\description{
\code{star_to_mat} returns a count matrix from STAR ReadsPerGene.out.tab
files.
}
\details{
This function will read in all the ReadsPerGene.out.tab files in the
specified directory and converts it to a counts matrix.
}
