% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{methods}
\name{[,BbcSE,ANY,ANY,ANY-method}
\alias{[,BbcSE,ANY,ANY,ANY-method}
\title{Subsetting method for BbcSE}
\usage{
\S4method{[}{BbcSE,ANY,ANY,ANY}(x, i, j, drop = TRUE)
}
\description{
Subsetting method for BbcSE
}
