% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{methods}
\name{[<-,BbcSE,ANY,ANY,BbcSE-method}
\alias{[<-,BbcSE,ANY,ANY,BbcSE-method}
\title{Assigning subsets method for BbcSE}
\usage{
\S4method{[}{BbcSE,ANY,ANY,BbcSE}(x, i, j, ...) <- value
}
\description{
Assigning subsets method for BbcSE
}
