% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{BbcSE-class}
\alias{BbcSE-class}
\alias{.BbcSE}
\title{BbcSE, an extension of SummarizedExperiment for VAI BBC RNA-seq workflow}
\description{
In a \code{BbcSE} object, "counts" must be the first assay and must contain
non-negative values.
}
