% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{BbcSE}
\alias{BbcSE}
\title{Constructor for BbcSE}
\usage{
BbcSE(counts, granges, ...)
}
\arguments{
\item{counts}{A count matrix with sample names as column names and gene names
as row names.}

\item{granges}{GRanges or GRangesList object. Optional.}

\item{...}{Arguments for SummarizedExperiment.}
}
\value{
A BbcSE object (extension of RangedSummarizedExperiment).
}
\description{
\code{BbcSE} is a constructor for BbcSE.
}
\details{
If GRanges object is provided, the names must match the rownames of the
    counts matrix. The output object will show only the intersected genes
    between the the counts and the GRanges.
}
