% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_data_methods.R
\name{ens2entrez}
\alias{ens2entrez}
\title{Convert Ensembl IDs to Entrez IDs for gene enrichment analyses.}
\usage{
ens2entrez(x, orgdb)
}
\arguments{
\item{x}{A BbcSE object.}

\item{orgdb}{A OrgDb object. Download the annotations for your species from
'http://bioconductor.org/packages/release/BiocViews.html#___OrgDb'}
}
\value{
A BbcSE object.
}
\description{
Convert Ensembl IDs to Entrez IDs for gene enrichment analyses.
}
\details{
Non 1:1 matches are determined from all the Ensembl IDs in the OrgDb, not
just those in the BbcSE object.
\itemize{
\item Entrez IDs that match multiple Ensembl IDs are NA
\item The first Entrez ID will be returned for genes with multiple possible
Entrez IDs.
\item Genes with no Entrez ID or absent from the database will be
NA. }
}
