% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edger_methods.R
\name{makeDGEList}
\alias{makeDGEList}
\title{Make a DGEList object.}
\usage{
makeDGEList(x, group = NULL)
}
\arguments{
\item{x}{A BbcSE object.}

\item{group}{Name of a column from colData()}
}
\value{
A BbcSE object.
}
\description{
Make a DGEList object based on assay(object, "counts").
}
