% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edger_methods.R
\name{makeDGEList}
\alias{makeDGEList}
\title{Make a DGEList object.}
\usage{
makeDGEList(x, group = NULL, rm_low_genes = TRUE, calc_norm = TRUE)
}
\arguments{
\item{x}{A BbcSE object.}

\item{group}{Name of a column from colData()}

\item{rm_low_genes}{logical indicating whether low count genes should be
removed. Implements edgeR::filterByExpr.}

\item{calc_norm}{logical indicating whether normalization factors should be
calculated. Implements edgeR::calcNormFactors. Also turns on calculations
of normalized counts and per-group normalized counts}
}
\value{
A BbcSE object.
}
\description{
Make a DGEList object based on assay(object, "counts").
}
\seealso{
\code{\link[edgeR]{filterByExpr}}
}
