% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{plot_PCA}
\alias{plot_PCA}
\title{Make PCA and scree plots}
\usage{
plot_PCA(x, norm_cts_type = "edger", color_by, shape_by)
}
\arguments{
\item{x}{A BbcSE object.}

\item{norm_cts_type}{"edger" or "deseq2"}

\item{color_by}{colData column to color by for PCA plot}

\item{shape_by}{colData column to shape by for PCA plot}
}
\value{
A list containing ggplot objects for 1. PCA 2. scree plot.
}
\description{
Make PCA and scree plots.
}
