% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BbcSE_manip_methods.R
\name{rbind,BbcSE-method}
\alias{rbind,BbcSE-method}
\title{Row combining method for BbcSE}
\usage{
\S4method{rbind}{BbcSE}(..., deparse.level = 1)
}
\arguments{
\item{...}{BbcSE objects}

\item{deparse.level}{see ?`rbind,SummarizedExperiment-method`}
}
\description{
Row combining method for BbcSE
}
